refactor(folder): component 하위에 페이지 폴더 생성
