+++ package.json
@@ -16,6 +16,7 @@
     "react-error-boundary": "^4.0.13",
     "react-router-dom": "6.23.1",
     "recoil": "^0.7.7",
+    "rian-modal-components": "^0.0.21",
     "styled-components": "^6.1.11"
   },
   "devDependencies": {
