+++ src/components/cart/PriceSection/PriceSection.skeleton.tsx
@@ -1,0 +1,13 @@
+import * as S from "./PriceSection.styles.ts";
+
+export const PriceSectionSkeleton = () => {
+    return (
+        <S.Wrapper>
+            <S.SkPriceInfoBox/>
+            <S.SkPriceInfoBox/>
+            <S.SkPriceInfoBox/>
+        </S.Wrapper>
+    );
+};
+
+export default PriceSectionSkeleton;
