+++ src/components/cart/ProductList/ProductList.skeleton.tsx
@@ -1,0 +1,15 @@
+import * as S from "./ProductList.style.ts";
+
+import ProductItemSkeleton from "../ProductItem/ProductItem.skeleton.tsx";
+
+const ProductListSkeleton = () => {
+    return (
+        <S.ListWrapper>
+            {Array.from({length: 3}).map((_, index) => (
+                <ProductItemSkeleton key={index}/>
+            ))}
+        </S.ListWrapper>
+    );
+};
+
+export default ProductListSkeleton;
