+++ src/pages/CartPage/CartPage.skeleton.tsx
@@ -1,20 +1,20 @@
-import PriceSectionSkeleton from "@/components/PriceSection/PriceSection.skeleton";
+import PriceSectionSkeleton from "@/components/cart/PriceSection/PriceSection.skeleton";
 import * as S from "./CartPage.style";
-import SkeletonProductList from "@/components/ProductList/ProductList.skeleton";
+import SkeletonProductList from "@/components/cart/ProductList/ProductList.skeleton";
 import SkTitleSet from "@/components/_common/TitleSet/TitleSet.skeleton";
 
 const CartPageSkeleton = () => {
-  return (
-    <S.CartItemListWrapper>
-      <SkTitleSet />
-      <S.SkCheckBoxWrapper />
-      <S.CheckBoxWrapper>
-        <S.SkCheckBoxWrapper />
-      </S.CheckBoxWrapper>
-      <SkeletonProductList />
-      <PriceSectionSkeleton />
-    </S.CartItemListWrapper>
-  );
+    return (
+        <S.CartItemListWrapper>
+            <SkTitleSet/>
+            <S.SkCheckBoxWrapper/>
+            <S.CheckBoxWrapper>
+                <S.SkCheckBoxWrapper/>
+            </S.CheckBoxWrapper>
+            <SkeletonProductList/>
+            <PriceSectionSkeleton/>
+        </S.CartItemListWrapper>
+    );
 };
 
 export default CartPageSkeleton;
